+++ internal/server/restapi/handler_schedules.go
@@ -284,7 +284,7 @@ func buildSchedule(info scheduler.JobInfo, row *ent.ScheduledJob) Schedule {
 		out.Paused = row.Paused
 		out.Status = ScheduleStatus(row.LastStatus)
 		out.LastFinishedAt = row.LastFinishedAt
-		out.LastDurationMs = int32(row.LastDurationMs)
+		out.LastDurationMs = row.LastDurationMs
 		if row.LastError != "" {
 			msg := row.LastError
 			out.LastError = &msg
